+++ CHANGES.rst
@@ -255,6 +255,9 @@ API changes
 
 - ``astropy.wcs``
 
+  - WCS objects can now be initialized with an ImageHDU or
+    PrimaryHDU object. [#4493]
+
 Bug fixes
 ^^^^^^^^^